test: move coverage tail tests into test_bmi_visualization_spec.py
- Move defensive fallback branch tests from separate file into main test file
- Ensures CI diff-cover sees these tests (covers lines 258, 361)
- Delete test_bmi_engine_breakpoints_coverage_tail.py to avoid confusion
- All coverage tail tests now in one place for better maintainability
